db_manager/requirements.txt
timer
mysql.connector

